refactor list.validateInput
