+++ admin/server/api/list/create.js
@@ -1,14 +1,4 @@
 var keystone = require('../../../../');
-var async = require('async');
-
-function getValidationError(path, msg, type) {
-	return {
-		name: 'ValidatorError',
-		path: path,
-		message: msg,
-		type: type || 'required'
-	};
-};
 
 module.exports = function(req, res) {
 	if (!keystone.security.csrf.validate(req)) {
@@ -24,20 +14,10 @@ module.exports = function(req, res) {
 	var files = req.files;
 	var item = new req.list.model();
 
-	var validationErrors = {};
-	async.forEach(req.list.fields, function(field, callback) {
-		// Validate the input for each field
-		field.validateInput(data, field.required, item,
-			function(err, isValid) {
-				if (!isValid) {
-					// Create object of validation errors
-					var errorMessage = 'Error: invalid input for ' + field.label;
-					var errorObject = getValidationError(field.path, errorMessage);
-					validationErrors[field.path] = errorObject;
-				}
-				callback();
-		});
-	}, function(err) {
+	req.list.validateInput(item, {
+		data: data,
+		files: files,
+	}, function(err, validationErrors) {
 		if (Object.keys(validationErrors).length > 0) {
 			// There were validation errors
 			res.status(500).json({
