+++ lib/list.js
@@ -176,6 +176,7 @@ List.prototype.relationship = require('./list/relationship');
 List.prototype.selectColumns = require('./list/selectColumns');
 List.prototype.updateItem = require('./list/updateItem');
 List.prototype.underscoreMethod = require('./list/underscoreMethod');
+List.prototype.validateInput = require('./list/validateInput');
 
 /*!
  * Export class
